+++ FPE/plugins/fileannouncer.py
@@ -11,7 +11,7 @@ class FileAnnouncerHandler(Handler):
     """File Announcer
     """
 
-    def __init__(self, handler_config) -> None:
+    def __init__(self, handler_config: dict[str, any]) -> None:
         """Initialise handler attributes.
         """
         self.handler_config = handler_config.copy()
@@ -23,7 +23,7 @@ class FileAnnouncerHandler(Handler):
             logging.info(
                 "File %s.", event.src_path)
 
-            if self.handler_config["deletesource"]:
+            if self.handler_config["deletesource"] and not os.path.isdir(event.src_path):
                 os.remove(event.src_path)
 
         except OSError as error:
